completed/backend/dna_mapper_lambda/requirements.txt
snowflake-connector-python==3.6.0
boto3==1.34.0
botocore==1.34.0
requests==2.31.0
pytz==2023.3 